Updated overwrite
Download trees that have not been downloaded, this allows it to be run
and paused without losing data

+++ stages/download.R
@@ -13,6 +13,8 @@ source (file.path ('tools', 'download_tools.R'))
 ## Parameters
 if (!exists ('min.taxa')) {
   min.taxa <- 100
+  subsample <- 100
+  overwrite <- FALSE
 }
 
 ## Dirs
@@ -25,7 +27,10 @@ if (!file.exists (output.dir)) {
 # First clear out all treebase trees from dir
 treebase.files <- list.files (path = output.dir,
                               pattern = '\\.tre')
-file.remove (file.path (output.dir, treebase.files))
+if (overwrite) {
+  file.remove (file.path (output.dir, treebase.files))
+  treebase.files <- NULL
+}
 # setup download log
 download.log <- file.path (output.dir, 'metadata.csv')
 if (file.exists (download.log)) {
@@ -48,6 +53,13 @@ cat (paste0 ('\n[', nrow (meta), '] matching trees in TreeBase ....'))
 cat (paste0 ('\n.... min size [', min (meta$ntaxa), ']'))
 cat (paste0 ('\n.... mean size [', round (mean (meta$ntaxa)), ']'))
 cat (paste0 ('\n.... max size [', max (meta$ntaxa), ']'))
+# if subsample, choose a random set of trees in meta
+if (is.numeric (subsample)) {
+  if (subsample < nrow (meta)) {
+    cat (paste0 ('\n.... taking [', subsample, '] subsample'))
+    meta <- meta[sample (1:nrow (meta), subsample), ]
+  }
+}
 
 ## Download
 cat ('\nDownloading trees ....')
@@ -55,13 +67,17 @@ counter <- 0
 for (i in 1:nrow (meta)) {
   cat (paste0 ('\n.... tree [',i,']\n'))
   tree.data <- meta[i, ]
+  filename <- paste0 (tree.data$Study.id, '.tre')
+  # only download tree if it hasn't already been downloaded
+  if (filename %in% treebase.files) {
+    next
+  }
   tree <- safeConnect (expr = {
     suppressWarnings (search_treebase (tree.data$Tree.id,
                                        by = 'id.tree',
                                        verbose = FALSE))})
   closeAllConnections ()
   if (length (tree) > 0) {
-    filename <- paste0 (tree.data$Study.id, '.tre')
     tree.data <- cbind (filename, tree.data)
     # control for malformed downloaded trees
     error <- try (write.tree (
